7.1. Refactoring.
