+++ src/util/const.js
@@ -4,6 +4,12 @@ const ActionType = {
   DELETE: `Delete`,
 };
 
+/**
+ * `Patch` is used to update only point
+ * `Minor` is used to update point, trip info(cost & start-middle-final point), filter view
+ * `Major` is used to update point, trip info, filter & `point board`
+ * @type {{MAJOR: string, MINOR: string, PATCH: string}}
+ */
 const UpdateType = {
   PATCH: `Patch`,
   MINOR: `Minor`,
