+++ src/util/point.js
