+++ src/view/abstract-view.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils/render";
+import {createElement} from "../util/render";
 
 class AbstractView {
   constructor() {
